tool_probe: tolerance window–based result selection
Previously, if you had SAMPLES_TOLERANCE set to a value, and SAMPLES set
to a number of probes, if you had SAMPLES probes within
SAMPLES_TOLERANCE, then those results would get poked into the median or
average function to return the result.  If there was a measurement
outside of that window, previous data would be discarded until there
were SAMPLES measurements in a row within SAMPLES_TOLERANCE.

     Probe samples exceed tolerance, retrying...

This new algorithm preserves this behavior for the initial SAMPLES worth
of probes; if they are all within SAMPLES_TOLERANCE, the result will
still be passed to median or average and the result yielded.  The
difference is what happens to measurements that _don't_ fall within
SAMPLES_TOLERANCE.  Instead of discarding these values, they are
retained, and a different method is used to look for a distinct 'peak'
in the values, at least SAMPLES higher than other potential ranges.

     Found 2 results at -0.268875 and -0.267312; but also 1 result at -0.175125; need more data...

This is done using a window function over the probe samples, and the
highest count window is chosen as a potential 'winner'.  Each of the
next most numerous windows are then compared, to see if any of them
would contradict the 'winner'.  If they are a distinctly different
range, that is, starting points are more than SAMPLES_TOLERANCE away,
and the number of measurements in the window is within SAMPLES of the
'winner', this is considered a contradiction, and probing continues.

     Found 9 results between -0.259813 and -0.252625; but also 7 results
between -0.272937 and -0.263250; need more data...

The probing will complete when a winner is found, _or_
SAMPLES_TOLERANCE_RETRIES * SAMPLES measurements (the previous worst,
pathological limit on the number of probes in the run) happen without a
clear winner being found.

     Peak in data found: 11 results between -0.259813 and -0.252625
     Result is z=-0.255125

The median value with this very tightly grouped set of results turns out
to be fairly stable indeed.

     Z is currently 1.74, net probe result is 0.02

This filter is designed for Maxwell coupling–based probes, which can
slightly degrade over time and start to return results where the pins
disconnected in a slightly different displacement orientation than
typical, for reasons yet to be clearly identified and probably relating
to creep and friction.  However, the approach should be appropriate for
many types of probes which have similar variations in their measured
values.  This could mean that people who previously used higher
tolerance values for more reliable probing, can return to using their
preferred tolerances.