tool_probe: widen tolerance if no result is found
If the tolerance is set tighter than the accuracy of the probe, then the
probing will continue until the maximum sample count has been read, and then
fail with:

     Results too scattered even after N sample(s): a, b, c, ...

However, there has often been a lot of data collected, and there may be a
signal within it.  Using a window for grouping values equal to the "tolerance"
may be fussier than needed.  It's quite probable that the median value of the
peak with a slightly wider window will still yield a value that should be
expected to be within the specified tolerance.  As such it's not really a
tolerance, but more of a granularity for the peak analysis.

If we reach the maximum number of samples, instead of failing and remaining
uncalibrated (which is arguably worse than giving up), progressively widen the
window starting from the specified tolerance and ending at 10x the tolerance.
Use whichever level first yields a clear answer: window at least 'SAMPLES' more
values in it than any other non–overlapping window.

The specifics of the window expasion should proably be configurable, but for
now these defaults seem reasonable.